check for script directory, exclude examples and .git directory

+++ run.sh
@@ -28,13 +28,21 @@ run_command_on_server() {
 echo "run command on server"
 ssh -o StrictHostKeyChecking=no -i $PEM_FILE ${USER}@${IP} << 'SCRIPT'
   cd /tmp/pscript;
-  mkdir /tmp/pscript;
+  if [ -e "/tmp/pscript" ]
+    then
+    echo "script directory exists"
+  else
+    mkdir /tmp/pscript;
+  fi
 SCRIPT
 }
 
 copy_to_server(){
-  rm provision_script_server.tar
-  tar -cvf  provision_script_server.tar . --exclude='servers'
+  if [ -e "rm provision_script_server.tar" ]
+    then
+     rm provision_script_server.tar
+  fi
+  tar -cvf  provision_script_server.tar . --exclude='servers'  --exclude='examples'  --exclude='.git'
   scp -o StrictHostKeyChecking=no -i $PEM_FILE  provision_script_server.tar ${USER}@${IP}:/tmp/pscript
 }
